Move completion base part into api
