+++ tests/end2end/features/completion/test_completion_bdd.py
@@ -8,7 +8,8 @@ import pytest
 import pytest_bdd as bdd
 
 from vimiv import api
-from vimiv.completion import completionmodels, completer
+from vimiv.apimodules import completer
+from vimiv.completion import completionmodels
 from vimiv.gui import completionwidget
